+++ battlecode22-scaffold/src/AFoolsGoldBot/RobotPlayer.java
@@ -21,7 +21,7 @@ public strictfp class RobotPlayer {
         // if (rc.getType() == RobotType.MINER) 
             // System.out.println(" Byte code C: " + Clock.getBytecodesLeft());
         while (true) {
-            // if(rc.getRoundNum() > 23) return;
+            // if(rc.getRoundNum() > 18) return;
             Globals.updateGlobals();
             try {
                 switch (rc.getType()) {