+++ battlecode22-scaffold/CHANGELOG.md
@@ -1,7 +1,8 @@
-# FirstPageBot
+# PageOneBot
 * Implement archon fleeing, may tweak value and other stuff regulating when it flees
 * Soldiers will now update comms with enemy archon location
 
 # FoolsGoldBot
 * Reintroduced update_enemy_archon location for soldiers
 * Some tweaking to the retreat function calculation
+* More accurate enemy locations and skip our own archon locations for small maps